Made changes to the phone number table etc.
Added a phone number id, as I realized it is possible for two different
families to be using the same phone number (they work at the same
business, for instance).  Created an extension field that is optional
and change phone number to a text field.   Also added a field for
tracking which phone number is the preferred one for a given family.

Added a phone type select to the add-record page and made changes to
charitydb.js to provide a list of phone types to populate that select.